# apis/open-api/iam/sdkwork-iam-open-api.openapi.yaml
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork IAM Open API",
    "version": "1.0.0",
    "description": "Public OAuth authorization-server ingress and provider callback contract for sdkwork-iam.",
    "x-sdkwork-api-authority": "sdkwork-iam-open-api",
    "x-sdkwork-sdk-family": "sdkwork-iam-open-sdk",
    "x-sdkwork-audience": "Public OAuth authorization-server clients, provider callbacks, and open ingress integrations"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local sdkwork-iam runtime"
    }
  ],
...
  "security": [],
  "paths": {
    "/.well-known/oauth-authorization-server": {
      "get": {
        "tags": [
          "iamOauth"
        ],
        "summary": "Iam oauth well Known authorization Server Metadata retrieve.",
        "operationId": "iam.oauth.wellKnown.authorizationServerMetadata.retrieve",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Upstream OAuth/OIDC discovery metadata document."
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [],
        "x-sdkwork-auth-mode": "compatibility",
        "x-sdkwork-forbid-credential-headers": false,
        "x-sdkwork-owner": "sdkwork-iam",
        "x-sdkwork-api-authority": "sdkwork-iam-open-api",
        "x-sdkwork-domain": "iam",
        "x-sdkwork-resource": "iam.oauth.wellKnown.authorizationServerMetadata",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "iam.oauth.read",
        "x-sdkwork-wire-protocol": "external",
        "x-sdkwork-external-protocol-id": "oauth-authorization-server-metadata"
      }
    },
    "/.well-known/openid-configuration": {
      "get": {
        "tags": [
          "iamOauth"
        ],
        "summary": "Iam oauth well Known openid Configuration retrieve.",
        "operationId": "iam.oauth.wellKnown.openidConfiguration.retrieve",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Upstream OAuth/OIDC discovery metadata document."
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [],
        "x-sdkwork-auth-mode": "compatibility",
        "x-sdkwork-forbid-credential-headers": false,
        "x-sdkwork-owner": "sdkwork-iam",
        "x-sdkwork-api-authority": "sdkwork-iam-open-api",
        "x-sdkwork-domain": "iam",
        "x-sdkwork-resource": "iam.oauth.wellKnown.openidConfiguration",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "iam.oauth.read",
        "x-sdkwork-wire-protocol": "external",
        "x-sdkwork-external-protocol-id": "oidc-discovery"
      }
    },
    "/iam/v3/api/oauth/authorize": {
      "get": {
        "tags": [
          "iamOauth"
        ],
        "summary": "Iam oauth authorize handle Get.",
        "operationId": "iam.oauth.authorize.handleGet",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [],
        "x-sdkwork-auth-mode": "anonymous",
        "x-sdkwork-forbid-credential-headers": false,
        "x-sdkwork-owner": "sdkwork-iam",
        "x-sdkwork-api-authority": "sdkwork-iam-open-api",
        "x-sdkwork-domain": "iam",
        "x-sdkwork-resource": "iam.oauth.authorize",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "iam.oauth.manage",
        "x-sdkwork-wire-protocol": "external",
        "x-sdkwork-external-protocol-id": "oauth-authorization-code"
      }
    },
    "/iam/v3/api/oauth/introspect": {
      "post": {
        "tags": [
          "iamOauth"
        ],
        "summary": "Iam oauth introspect create.",
        "operationId": "iam.oauth.introspect.create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [],
        "x-sdkwork-auth-mode": "anonymous",
        "x-sdkwork-forbid-credential-headers": false,
        "x-sdkwork-owner": "sdkwork-iam",
        "x-sdkwork-api-authority": "sdkwork-iam-open-api",
        "x-sdkwork-domain": "iam",
        "x-sdkwork-resource": "iam.oauth.introspect",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "iam.oauth.manage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppbaseOperationCommand"
              }
            }
          }
        },
        "x-sdkwork-wire-protocol": "external",
        "x-sdkwork-external-protocol-id": "oauth-introspect"
      }
    },
    "/iam/v3/api/oauth/jwks": {
      "get": {
        "tags": [
          "iamOauth"
        ],
        "summary": "Iam oauth jwks retrieve.",
        "operationId": "iam.oauth.jwks.retrieve",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [],
        "x-sdkwork-auth-mode": "anonymous",
        "x-sdkwork-forbid-credential-headers": false,
        "x-sdkwork-owner": "sdkwork-iam",
        "x-sdkwork-api-authority": "sdkwork-iam-open-api",
        "x-sdkwork-domain": "iam",
        "x-sdkwork-resource": "iam.oauth.jwks",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "iam.oauth.read",
        "x-sdkwork-wire-protocol": "external",
        "x-sdkwork-external-protocol-id": "oauth-jwks"
      }
    },
...
    "/iam/v3/api/oauth/revoke": {
      "post": {
        "tags": [
          "iamOauth"
        ],
        "summary": "Iam oauth revoke create.",
        "operationId": "iam.oauth.revoke.create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkResourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [],
        "x-sdkwork-auth-mode": "anonymous",
        "x-sdkwork-forbid-credential-headers": false,
        "x-sdkwork-owner": "sdkwork-iam",
        "x-sdkwork-api-authority": "sdkwork-iam-open-api",
        "x-sdkwork-domain": "iam",
        "x-sdkwork-resource": "iam.oauth.revoke",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-server-request-id": true,
        "x-sdkwork-permission": "iam.oauth.manage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppbaseOperationCommand"
              }
            }
          }
        },
        "x-sdkwork-wire-protocol": "external",
        "x-sdkwork-external-protocol-id": "oauth-revoke"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "SDKWork open-api API key for protected ingress operations."
      },
      "OAuthBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OAuth bearer credential for protected open-api operations."
      }
    },
    "schemas": {
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
      "SdkWorkResourceData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "item"
        ],
        "properties": {
          "item": {
            "type": "object",
            "additionalProperties": true,
            "description": "Typed domain resource for the operation."
          }
        }
      },
...
      "SdkWorkPlatformErrorCode": {
        "type": "integer",
        "format": "int32",
        "minimum": 40001,
        "maximum": 79999,
        "description": "Platform or domain error code per API_SPEC.md §15.3."
      },
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "i18nKey": {
            "type": "string",
            "description": "Optional stable localization key such as errors.result.40001."
          },
          "locale": {
            "type": "string",
            "description": "Optional effective BCP 47 locale used by framework message mapping."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32",
            "minimum": 40011,
            "maximum": 40099
          },
          "i18nKey": {
            "type": "string"
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                }
              ]
            }
          }
        }
      },
      "SdkWorkResourceResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SdkWorkResourceData"
              }
            }
          }
        ]
      },
...
      "AppbaseOperationCommand": {
        "type": "object",
        "additionalProperties": true,
        "description": "Operation-specific command payload defined by the owning IAM Rust module."
      },
...
    }
  },
...
}